requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Machine learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2

# Environment variables
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1